backend/requirements.txt
# --- Ajan: MIMAR (THE ARCHITECT) ---
# Python backend bagimliklar

fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiomysql==0.2.0
alembic==1.14.0
redis[hiredis]==5.2.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
websockets==13.1
httpx==0.28.0
PyJWT==2.9.0
cryptography==44.0.0
python-multipart==0.0.18
bcrypt==4.2.0
